winio: Add support for WINIO to process.
